+++ code/src/ui/graphic/table/registry/registry-table.tsx
@@ -3,7 +3,7 @@ import styles from './registry.table.module.css';
 import React from 'react';
 import { FieldValues } from 'react-hook-form';
 
-import { RegistryFieldValues } from 'types/form';
+import { RegistryFieldValues, RegistryTaskOption } from 'types/form';
 import { parseWordsForLabels } from 'utils/client-utils';
 import RegistryRowActions from './actions/registry-table-action';
 import StatusComponent from 'ui/text/status/status';
@@ -14,10 +14,9 @@ import { RegistryTableTheme } from './registry-table-theme';
 
 interface RegistryTableProps {
   recordType: string;
-  isTaskPage: boolean;
+  lifecycleStage: string;
   instances: RegistryFieldValues[];
-  setTaskId: React.Dispatch<React.SetStateAction<string>>;
-  setTaskStatus: React.Dispatch<React.SetStateAction<string>>;
+  setTask: React.Dispatch<React.SetStateAction<RegistryTaskOption>>;
   limit?: number;
 }
 
@@ -25,10 +24,9 @@ interface RegistryTableProps {
  * This component renders a registry of table based on the inputs.
  * 
  * @param {string} recordType The type of the record.
- * @param {boolean} isTaskPage Indicator if the table is currently on the task view.
+ * @param {string} lifecycleStage The current stage of a contract lifecycle to display.
  * @param {RegistryFieldValues[]} instances The instance values for the table.
- * @param setTaskId A dispatch method to set task id when required.
- * @param setTaskStatus A dispatch method to set task status when required.
+ * @param setTask A dispatch method to set the task option when required.
  * @param {number} limit Optional limit to the number of columns shown.
  */
 export default function RegistryTable(props: Readonly<RegistryTableProps>) {
@@ -44,14 +42,18 @@ export default function RegistryTable(props: Readonly<RegistryTableProps>) {
         renderCell: (params) => {
           return (<RegistryRowActions
             recordType={props.recordType}
-            isTaskPage={props.isTaskPage}
+            lifecycleStage={props.lifecycleStage}
             row={params.row}
-            setTaskId={props.setTaskId}
-            setTaskStatus={props.setTaskStatus}
+            setTask={props.setTask}
           />);
         }
       },
-      ...Object.keys(props.instances[0]).map(field => ({
+      // Get instances with the most number of fields
+      ...Object.keys(props.instances.reduce((prev, current) => {
+        const prevKeys = Object.keys(prev).length;
+        const currentKeys = Object.keys(current).length;
+        return prevKeys >= currentKeys ? prev : current;
+      })).map(field => ({
         field,
         headerName: parseWordsForLabels(field),
         width: 150, // Adjust the width as needed